+++ src/Resolver/Resolver.php
@@ -4,7 +4,7 @@ namespace Icicle\Dns\Resolver;
 use Icicle\Coroutine\Coroutine;
 use Icicle\Dns\Exception\NotFoundException;
 use Icicle\Dns\Executor\ExecutorInterface;
-use Icicle\Promise\Promise;
+use Icicle\Promise;
 use LibDNS\Records\ResourceTypes;
 
 class Resolver implements ResolverInterface
@@ -31,23 +31,10 @@ class Resolver implements ResolverInterface
         $retries = ExecutorInterface::DEFAULT_RETRIES
     ) {
         if (strtolower($domain) === 'localhost') {
-            return Promise::resolve(['127.0.0.1']);
+            yield ['127.0.0.1'];
+            return;
         }
 
-        return new Coroutine($this->run($domain, $timeout, $retries));
-    }
-
-    /**
-     * @coroutine
-     *
-     * @param   string $domain
-     * @param   float|int $timeout
-     * @param   int $retries
-     *
-     * @return  \Generator
-     */
-    protected function run($domain, $timeout, $retries)
-    {
         /** @var \LibDNS\Messages\Message $response */
         $response = (yield $this->executor->execute($domain, ResourceTypes::A, $timeout, $retries));