+++ tests/Executor/MultiExecutorTest.php
@@ -1,10 +1,11 @@
 <?php
 namespace Icicle\Tests\Dns\Executor;
 
+use Icicle\Coroutine\Coroutine;
 use Icicle\Dns\Exception\MessageException;
 use Icicle\Dns\Executor\MultiExecutor;
-use Icicle\Loop\Loop;
-use Icicle\Promise\Promise;
+use Icicle\Loop;
+use Icicle\Promise;
 use Icicle\Tests\Dns\TestCase;
 
 class MultiExecutorTest extends TestCase
@@ -29,15 +30,15 @@ class MultiExecutorTest extends TestCase
 
     public function testNoExecutors()
     {
-        $promise = $this->executor->execute('example.com', 'A');
+        $coroutine = new Coroutine($this->executor->execute('example.com', 'A'));
 
         $callback = $this->createCallback(1);
         $callback->method('__invoke')
             ->with($this->isInstanceOf('Icicle\Dns\Exception\LogicException'));
 
-        $promise->done($this->createCallback(0), $callback);
+        $coroutine->done($this->createCallback(0), $callback);
 
-        Loop::run();
+        Loop\run();
     }
 
     public function execute($type, $domain, $request, $response, array $answers = null, array $authority = null)
@@ -49,19 +50,19 @@ class MultiExecutorTest extends TestCase
         $executor->expects($this->once())
             ->method('execute')
             ->with($domain, $type)
-            ->will($this->returnValue(Promise::resolve($message)));
+            ->will($this->returnValue(Promise\resolve($message)));
 
         $this->executor->add($executor);
 
-        $promise = $this->executor->execute($domain, $type);
+        $coroutine = new Coroutine($this->executor->execute($domain, $type));
 
         $callback = $this->createCallback(1);
         $callback->method('__invoke')
             ->with($this->identicalTo($message));
 
-        $promise->done($callback, $this->createCallback(0));
+        $coroutine->done($callback, $this->createCallback(0));
 
-        Loop::run();
+        Loop\run();
     }
 
     /**
@@ -94,7 +95,7 @@ class MultiExecutorTest extends TestCase
 
         $executor->expects($this->once())
             ->method('execute')
-            ->will($this->returnValue(Promise::reject(new MessageException())));
+            ->will($this->returnValue(Promise\reject(new MessageException())));
 
         $this->executor->add($executor);
 
@@ -102,19 +103,19 @@ class MultiExecutorTest extends TestCase
 
         $executor->expects($this->once())
             ->method('execute')
-            ->will($this->returnValue(Promise::resolve($this->createMessage())));
+            ->will($this->returnValue(Promise\resolve($this->createMessage())));
 
         $this->executor->add($executor);
 
-        $promise = $this->executor->execute('example.com', 'A');
+        $coroutine = new Coroutine($this->executor->execute('example.com', 'A'));
 
         $callback = $this->createCallback(1);
         $callback->method('__invoke')
             ->with($this->isInstanceOf('LibDNS\Messages\Message'));
 
-        $promise->done($callback, $this->createCallback(0));
+        $coroutine->done($callback, $this->createCallback(0));
 
-        Loop::run();
+        Loop\run();
     }
 
     public function testNextRequestUsesLastRespondingExecutor()
@@ -123,7 +124,7 @@ class MultiExecutorTest extends TestCase
 
         $executor->expects($this->once())
             ->method('execute')
-            ->will($this->returnValue(Promise::reject(new MessageException())));
+            ->will($this->returnValue(Promise\reject(new MessageException())));
 
         $this->executor->add($executor);
 
@@ -131,29 +132,29 @@ class MultiExecutorTest extends TestCase
 
         $executor->expects($this->exactly(2))
             ->method('execute')
-            ->will($this->returnValue(Promise::resolve($this->createMessage())));
+            ->will($this->returnValue(Promise\resolve($this->createMessage())));
 
         $this->executor->add($executor);
 
-        $promise = $this->executor->execute('example.com', 'A');
+        $coroutine = new Coroutine($this->executor->execute('example.com', 'A'));
 
         $callback = $this->createCallback(1);
         $callback->method('__invoke')
             ->with($this->isInstanceOf('LibDNS\Messages\Message'));
 
-        $promise->done($callback, $this->createCallback(0));
+        $coroutine->done($callback, $this->createCallback(0));
 
-        Loop::run(); // Should shift first executor to back of list.
+        Loop\run(); // Should shift first executor to back of list.
 
-        $promise = $this->executor->execute('example.org', 'A');
+        $coroutine = new Coroutine($this->executor->execute('example.org', 'A'));
 
         $callback = $this->createCallback(1);
         $callback->method('__invoke')
             ->with($this->isInstanceOf('LibDNS\Messages\Message'));
 
-        $promise->done($callback, $this->createCallback(0));
+        $coroutine->done($callback, $this->createCallback(0));
 
-        Loop::run(); // Should call second executor first.
+        Loop\run(); // Should call second executor first.
     }
 
     public function testRetries()
@@ -166,7 +167,7 @@ class MultiExecutorTest extends TestCase
 
         $executor->expects($this->exactly($retries + 1))
             ->method('execute')
-            ->will($this->returnValue(Promise::reject($exception)));
+            ->will($this->returnValue(Promise\reject($exception)));
 
         $this->executor->add($executor);
 
@@ -174,19 +175,19 @@ class MultiExecutorTest extends TestCase
 
         $executor->expects($this->exactly($retries + 1))
             ->method('execute')
-            ->will($this->returnValue(Promise::reject($exception)));
+            ->will($this->returnValue(Promise\reject($exception)));
 
         $this->executor->add($executor);
 
-        $promise = $this->executor->execute('example.com', 'A', $timeout, $retries);
+        $coroutine = new Coroutine($this->executor->execute('example.com', 'A', $timeout, $retries));
 
         $callback = $this->createCallback(1);
         $callback->method('__invoke')
             ->with($this->identicalTo($exception));
 
-        $promise->done($this->createCallback(0), $callback);
+        $coroutine->done($this->createCallback(0), $callback);
 
-        Loop::run();
+        Loop\run();
     }
 
     /**
@@ -201,18 +202,18 @@ class MultiExecutorTest extends TestCase
 
         $executor->expects($this->once())
             ->method('execute')
-            ->will($this->returnValue(Promise::reject($exception)));
+            ->will($this->returnValue(Promise\reject($exception)));
 
         $this->executor->add($executor);
 
-        $promise = $this->executor->execute('example.com', 'A', $timeout, -1);
+        $coroutine = new Coroutine($this->executor->execute('example.com', 'A', $timeout, -1));
 
         $callback = $this->createCallback(1);
         $callback->method('__invoke')
             ->with($this->identicalTo($exception));
 
-        $promise->done($this->createCallback(0), $callback);
+        $coroutine->done($this->createCallback(0), $callback);
 
-        Loop::run();
+        Loop\run();
     }
 }