+++ tests/Resolver/ResolverTest.php
@@ -1,9 +1,10 @@
 <?php
 namespace Icicle\Tests\Dns\Resolver;
 
+use Icicle\Coroutine\Coroutine;
 use Icicle\Dns\Exception\NotFoundException;
 use Icicle\Dns\Resolver\Resolver;
-use Icicle\Loop\Loop;
+use Icicle\Loop;
 use Icicle\Tests\Dns\TestCase;
 use LibDNS\Records\ResourceQTypes;
 use Mockery;
@@ -44,7 +45,7 @@ class ResolverTest extends TestCase
             ->with(Mockery::mustBe($domain), Mockery::mustBe(ResourceQTypes::A), Mockery::any(), Mockery::type('integer'))
             ->andReturn($this->createMessage($answers, $authority));
 
-        $promise = $this->resolver->resolve($domain);
+        $coroutine = new Coroutine($this->resolver->resolve($domain));
 
         $result = [];
 
@@ -58,9 +59,9 @@ class ResolverTest extends TestCase
         $callback->method('__invoke')
             ->with($this->equalTo($result));
 
-        $promise->done($callback, $this->createCallback(0));
+        $coroutine->done($callback, $this->createCallback(0));
 
-        Loop::run();
+        Loop\run();
     }
 
     /**
@@ -73,7 +74,7 @@ class ResolverTest extends TestCase
         $this->executor->shouldReceive('execute')
             ->andReturn($this->createMessage());
 
-        $promise = $this->resolver->resolve($domain);
+        $coroutine = new Coroutine($this->resolver->resolve($domain));
 
         $callback = $this->createCallback(1);
         $callback->method('__invoke')
@@ -83,9 +84,9 @@ class ResolverTest extends TestCase
                 $this->assertSame(ResourceQTypes::A, $exception->getType());
             }));
 
-        $promise->done($this->createCallback(0), $callback);
+        $coroutine->done($this->createCallback(0), $callback);
 
-        Loop::run();
+        Loop\run();
     }
 
     public function testLocalhost()
@@ -93,14 +94,14 @@ class ResolverTest extends TestCase
         $this->executor->shouldReceive('execute')
             ->never();
 
-        $promise = $this->resolver->resolve('localhost');
+        $coroutine = new Coroutine($this->resolver->resolve('localhost'));
 
         $callback = $this->createCallback(1);
         $callback->method('__invoke')
             ->with($this->equalTo(['127.0.0.1']));
 
-        $promise->done($callback, $this->createCallback(0));
+        $coroutine->done($callback, $this->createCallback(0));
 
-        Loop::run();
+        Loop\run();
     }
 }